Financepeer_task/Backend_part/requirements.txt
asgiref==3.4.1
cffi==1.15.0
cryptography==36.0.1
Django==3.2
django-filter==21.1
django-rest-knox==4.1.0
djangorestframework==3.13.1
importlib-metadata==4.10.0
Markdown==3.3.6
pycparser==2.21
pytz==2021.3
sqlparse==0.4.2
zipp==3.7.0
django-cors-headers==3.11.0
